backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
psycopg[binary]>=3.1,<4
gunicorn>=21.0,<23
python-dotenv>=1.0,<2
pytest>=7.4,<8
pytest-django>=4.5,<5
pytest-cov>=4.1,<5
factory-boy>=3.3,<4
